pedantic clippy changes

+++ src/components.rs
@@ -31,9 +31,9 @@ pub struct ItemShape {
 
 #[derive(Clone, Copy)]
 pub struct Position {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 pub struct Hitbox {
